修改了：原料强制出库时报错的问题 修改了：金料和碎石出入库记录备注

+++ mzf/src/com/zonrong/inventory/service/MaterialInventoryService.java
@@ -1,19 +1,5 @@
 package com.zonrong.inventory.service;
 
-import java.math.BigDecimal;
-import java.math.MathContext;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-import javax.annotation.Resource;
-
-import org.apache.commons.collections.CollectionUtils;
-import org.apache.commons.collections.MapUtils;
-import org.apache.commons.lang.ArrayUtils;
-import org.apache.log4j.Logger;
-import org.springframework.stereotype.Service;
-
 import com.zonrong.basics.material.service.MaterialService;
 import com.zonrong.common.service.MzfOrgService;
 import com.zonrong.common.utils.MzfEntity;
@@ -24,6 +10,18 @@ import com.zonrong.inventory.service.InventoryService.BizType;
 import com.zonrong.inventory.service.InventoryService.InventoryType;
 import com.zonrong.metadata.EntityMetadata;
 import com.zonrong.metadata.service.MetadataProvider;
+import org.apache.commons.collections.CollectionUtils;
+import org.apache.commons.collections.MapUtils;
+import org.apache.commons.lang.ArrayUtils;
+import org.apache.log4j.Logger;
+import org.springframework.stereotype.Service;
+
+import javax.annotation.Resource;
+import java.math.BigDecimal;
+import java.math.MathContext;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
 /**
  * date: 2010-11-3
@@ -34,7 +32,7 @@ import com.zonrong.metadata.service.MetadataProvider;
 @Service
 public class MaterialInventoryService {
 	private Logger logger = Logger.getLogger(this.getClass());
-	
+
 	@Resource
 	private MetadataProvider metadataProvider;
 	@Resource
@@ -45,12 +43,12 @@ public class MaterialInventoryService {
 	private MaterialService materialService;
 	@Resource
 	private MzfOrgService mzfOrgService;
-	
+
 	public void warehouse(BizType bizType, int materialId, BigDecimal quantity, BigDecimal cost, String costDesc, String remark, IUser user) throws BusinessException {
 		int orgId = user.getOrgId();
 		warehouse( bizType, orgId, materialId, quantity, cost, costDesc, remark, user);
 	}
-	
+
 	public void warehouse(BizType bizType,int orgId, int materialId, BigDecimal quantity, BigDecimal cost, String costDesc, String remark, IUser user) throws BusinessException {
 		Map<String, Object> inventory = inventoryService.findMaterialInventory(materialId, orgId, user);
 		Integer inventoryId = null;
@@ -59,22 +57,22 @@ public class MaterialInventoryService {
 		} else {
 			inventoryId = MapUtils.getInteger(inventory, "id");
 		}
-		
-		inventoryService.addQuantity(bizType, inventoryId, quantity, cost, costDesc, remark, user);		
+
+		inventoryService.addQuantity(bizType, inventoryId, quantity, cost, costDesc, remark, user);
 	}
-	
-	
+
+
 	public void delivery(BizType bizType, int inventoryId, BigDecimal quantity, BigDecimal cost, String costDesc, boolean isUpdateLockedQuantity, String remark, IUser user) throws BusinessException {
 		inventoryService.deliveryByQuantityFlow(bizType, inventoryId, quantity, cost, costDesc, isUpdateLockedQuantity, remark, user);
-	} 	
-	
+	}
+
 	public void deliveryByQuantity(BizType bizType, int materialId, BigDecimal quantity, BigDecimal cost, String costDesc, int orgId, boolean isUpdateLockedQuantity, String remark, IUser user) throws BusinessException {
 		Map<String, Object> inventory = getMaterialInventory(materialId, orgId, user);
 		Integer inventoryId = MapUtils.getInteger(inventory, "inventoryId");
-		
+
 		delivery(bizType, inventoryId, quantity, cost, costDesc, isUpdateLockedQuantity, remark, user);
 	}
-	
+
 	public List<Map<String, Object>> listMaterialInventory(Integer[] materialIds, int orgId, IUser user) throws BusinessException {
 		EntityMetadata metadata = metadataProvider.getEntityMetadata(MzfEntity.MATERIAL_INVENTORY_VIEW);
 		Map<String, Object> where = new HashMap<String, Object>();
@@ -82,27 +80,27 @@ public class MaterialInventoryService {
 			where.put("id", materialIds);
 		}
 		where.put("orgId", orgId);
-		List<Map<String, Object>> list = entityService.list(metadata, where, null, user.asSystem());		
+		List<Map<String, Object>> list = entityService.list(metadata, where, null, user.asSystem());
 		return list;
 	}
-	
+
 	public Map<String, Object> getMaterialInventory(int materialId, int orgId, IUser user) throws BusinessException {
-		List<Map<String, Object>> list = listMaterialInventory(new Integer[]{materialId}, orgId, user);	
+		List<Map<String, Object>> list = listMaterialInventory(new Integer[]{materialId}, orgId, user);
 		if (CollectionUtils.isEmpty(list)) {
 			throw new BusinessException("库存中未找到该物料[" + materialId + "]");
 		} else if (list.size() > 1) {
 			throw new BusinessException("库存中找到多件物料[" + materialId + "]");
 		}
 		Map<String, Object> inventory = list.get(0);
-		
+
 		return inventory;
-	}	
-	
+	}
+
 //	public void lockByQuantityByInventoryId(Map<Integer, BigDecimal> map, String remark, IUser user) throws BusinessException {
 //		if (MapUtils.isEmpty(map)) {
 //			return;
 //		}
-//		
+//
 //		for (Iterator<Entry<Integer, BigDecimal>> it = map.entrySet().iterator(); it.hasNext();) {
 //			Entry<Integer, BigDecimal> en = it.next();
 //			int inventoryId = en.getKey();
@@ -110,13 +108,13 @@ public class MaterialInventoryService {
 //			inventoryService.addLockedQuantity(inventoryId, lockedQuantity, user);
 //		}
 //	}
-//	
+//
 //	public void lockByQuantity(Map<Integer, BigDecimal> map, String remark, IUser user) throws BusinessException {
 //		if (MapUtils.isEmpty(map)) {
 //			return;
 //		}
 //
-//		Map<Integer, BigDecimal> newMap = convert(map, user);		
+//		Map<Integer, BigDecimal> newMap = convert(map, user);
 //		for (Iterator<Entry<Integer, BigDecimal>> it = newMap.entrySet().iterator(); it.hasNext();) {
 //			Entry<Integer, BigDecimal> en = it.next();
 //			int inventoryId = en.getKey();
@@ -124,13 +122,13 @@ public class MaterialInventoryService {
 //			inventoryService.addLockedQuantity(inventoryId, lockedQuantity, user);
 //		}
 //	}
-	
+
 //	public void freeByQuantity(Map<Integer, BigDecimal> map, String remark, IUser user) throws BusinessException {
 //		if (MapUtils.isEmpty(map)) {
 //			return;
 //		}
 //
-//		Map<Integer, BigDecimal> newMap = convert(map, user);		
+//		Map<Integer, BigDecimal> newMap = convert(map, user);
 //		for (Iterator<Entry<Integer, BigDecimal>> it = newMap.entrySet().iterator(); it.hasNext();) {
 //			Entry<Integer, BigDecimal> en = it.next();
 //			int inventoryId = en.getKey();
@@ -139,14 +137,14 @@ public class MaterialInventoryService {
 //			inventoryService.addLockedQuantity(inventoryId, lockedQuantity, user);
 //		}
 //	}
-	
+
 	public void send(int materialId, BigDecimal quantity, int sourceOrgId, String remark, IUser user) throws BusinessException {
 		Map<String, Object> inventory =  getMaterialInventory(materialId, sourceOrgId, user);
 		Integer inventoryId = MapUtils.getInteger(inventory, "inventoryId");
 		BigDecimal cost = new BigDecimal(MapUtils.getString(inventory, "unitCost","0")).multiply(quantity,MathContext.DECIMAL32);
-		inventoryService.createFlowOnQuantity(BizType.send, inventoryId, quantity, InventoryType.delivery, cost, null, remark, user);
+		inventoryService.createFlowOnQuantity(BizType.send, inventoryId, quantity, InventoryType.delivery, cost.doubleValue(), null, remark, user);
 	}
-	
+
 	public void receive(int materialId, BigDecimal quantity, BigDecimal cost, int sourceOrgId, int targetOrgId, String remark, IUser user) throws BusinessException {
 		if (targetOrgId != user.getOrgId()) {
 			throw new BusinessException("操作员所在部门非调入部门，不允许收货");
@@ -154,17 +152,17 @@ public class MaterialInventoryService {
 		if (sourceOrgId == targetOrgId) {
 			throw new BusinessException("调出和调入部门相同，不允许收货");
 		}
-		
+
 		BigDecimal warehouseCost = cost;
 		if (mzfOrgService.isHq(sourceOrgId) || mzfOrgService.isHq(targetOrgId)) {
 			//更新总部物料成本
 			Map<String, Object> inventoryHQ =  getMaterialInventory(materialId, mzfOrgService.getHQOrgId(), user);
 			BigDecimal costHQ = new BigDecimal(MapUtils.getString(inventoryHQ, "cost", Integer.toString(0)));
 			BigDecimal quantityHQ = new BigDecimal(MapUtils.getString(inventoryHQ, "quantity", Integer.toString(0)));
-			
+
 			BigDecimal newCostHQ = null;
 			if (quantityHQ.doubleValue() != 0) {
-				if (mzfOrgService.isHq(sourceOrgId)) {					
+				if (mzfOrgService.isHq(sourceOrgId)) {
 					newCostHQ = (quantityHQ.subtract(quantity)).divide(quantityHQ, 2, BigDecimal.ROUND_HALF_EVEN).multiply(costHQ);
 				}
 				if (mzfOrgService.isHq(targetOrgId)) {
@@ -172,25 +170,25 @@ public class MaterialInventoryService {
 				}
 				warehouseCost = newCostHQ.subtract(costHQ);
 			} else {
-				newCostHQ = cost.abs();				
+				newCostHQ = cost.abs();
 			}
-			
+
 			Map<String, Object> field = new HashMap<String, Object>();
 			field.put("cost", newCostHQ);
 			entityService.updateById(MzfEntity.MATERIAL, Integer.toString(materialId), field, user);
 		} else {
 			throw new BusinessException("只允许总部与门店之间调拨");
 		}
-		
-		
+
+
 		Map<String, Object> inventory =  getMaterialInventory(materialId, sourceOrgId, user);
 		Integer inventoryId = MapUtils.getInteger(inventory, "inventoryId");
 		String costDesc = "批发价: " + cost;
 //		delivery(BizType.receive, inventoryId, quantity, cost, costDesc, true, remark, user);
-		
-		inventoryService.deliveryByQuantityIgnoreFlow(BizType.receive, inventoryId, quantity, null, null, true, remark, user);		
+
+		inventoryService.deliveryByQuantityIgnoreFlow(BizType.receive, inventoryId, quantity, null, null, true, remark, user);
 		warehouse(BizType.receive, materialId, quantity, warehouseCost.abs(), costDesc, remark, user);
-		
+
 //		materialService.addCost(materialId, cost.abs().multiply(new BigDecimal(-1)), sourceOrgId, user);
 //		materialService.addCost(materialId, cost.abs(), targetOrgId, user);
 	}
@@ -218,12 +216,12 @@ public class MaterialInventoryService {
 //			if (inventoryId == null || lockedQuantity == null) {
 //				throw new BusinessException("获取参数错误");
 //			}
-//			
+//
 //			newMap.put(inventoryId, lockedQuantity);
 //		}
-//		
-//		return newMap;		
-//	}	
+//
+//		return newMap;
+//	}
 }
